# internal/pricemon/repository/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"database/sql/driver"
	"fmt"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type OfferSide string

const (
	OfferSideBuy  OfferSide = "buy"
	OfferSideSell OfferSide = "sell"
)

func (e *OfferSide) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = OfferSide(s)
	case string:
		*e = OfferSide(s)
	default:
		return fmt.Errorf("unsupported scan type for OfferSide: %T", src)
	}
	return nil
}

type NullOfferSide struct {
	OfferSide OfferSide
	Valid     bool // Valid is true if OfferSide is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullOfferSide) Scan(value interface{}) error {
	if value == nil {
		ns.OfferSide, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.OfferSide.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullOfferSide) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.OfferSide), nil
}

type BestDeal struct {
	ItemID         int64
	BuyPlatformID  int64
	SellPlatformID int64
	BuyPrice       int64
	SellPrice      int64
	Profit         pgtype.Int8
	UpdatedAt      pgtype.Timestamptz
}

type Collector struct {
	ID         uuid.UUID
	PlatformID int64
	Name       string
	ApiKeyHash []byte
	CreatedAt  pgtype.Timestamptz
}

type Item struct {
	ID   int64
	Name string
}

type Offer struct {
	PlatformID int64
	ItemID     int64
	Side       OfferSide
	Price      int64
	UpdatedAt  pgtype.Timestamptz
}

type Platform struct {
	ID        int64
	Name      string
	CreatedAt pgtype.Timestamptz
}